% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/negate.R
\name{\%ni\%}
\alias{\%ni\%}
\title{Non-matching values}
\usage{
x \%ni\% y
}
\arguments{
\item{x}{vector or \code{NULL}}

\item{y}{vector or \code{NULL}}
}
\description{
This tests for non-membership of a given vector.
}
\examples{
1:10 \%ni\% c(1, 3, 5, 9)

}

