% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tardis.R
\name{parse_datetime}
\alias{parse_datetime}
\title{Parses a supplied date/time column to new columns of individual
temporal elements for additional manipulation.}
\usage{
parse_datetime(my.df, my.df.col)
}
\arguments{
\item{my.df}{Data frame}

\item{my.df.col}{Column with date time values of class POSIX.}
}
\value{
my.df. Essentially cbinds original data frame with temporal data.
}
\description{
Parses a supplied date/time column to new columns of individual
temporal elements for additional manipulation.
}
\seealso{
Other res: \code{\link{\%ni\%}}
}

