% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profile.R
\name{reload_profile}
\alias{reload_profile}
\title{Reloads the .Rprofile}
\usage{
reload_profile(my.dir = Sys.getenv("HOME"))
}
\description{
This is useful for cases when we clear our workspace, but would like to keep
variables we have defined within our profile.
}
\section{Warnings}{

Defaults to the global .Rprofile, but can be changed to load a different
.Rprofile (e.g. for a Rproject).

param my.dir Parent directory for the .Rprofile
}

